requirements.txt
plotly-express 
pandas 
numpy 
scikit-learn 
streamlit